feat(indian-farmer-assistance-app): Add admin service enhancements, API gateway implementation, and AI service optimizations
- Add audit logging aspect and interceptor for comprehensive request/response tracking
- Implement WebMvcConfig for cross-cutting concerns and request handling
- Add data retention service for managing audit log lifecycle
- Enhance AuditLog and Document entities with additional metadata fields
- Implement API gateway with authentication, authorization, and rate limiting filters
- Add gateway configuration, OpenAPI documentation, and fallback controller
- Create bandwidth service with adaptive quality and data usage monitoring
- Add GPU utilities and model caching for AI service performance optimization
- Implement disease detection model integration tests
- Add common utilities for error handling, crash recovery, and performance optimization
- Consolidate property-based tests into unified unit test suite
- Update application configurations for dev, staging, and production environments
- Add comprehensive documentation files for agriculture inputs and KRISHI portal
- Initialize .gitignore for project-wide version control management

+++ indian-farmer-assistance-app/backend/crop-service/src/main/java/com/farmer/crop/dto/ClimateRiskDto.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -40,7 +39,7 @@ public class ClimateRiskDto {
     /**
      * Risk score (0-100, higher = more risky)
      */
-    private BigDecimal riskScore;
+    private Double riskScore;
     
     /**
      * Rainfall deviation scenario analysis
@@ -109,13 +108,13 @@ public class ClimateRiskDto {
         /**
          * Historical average rainfall (mm)
          */
-        private BigDecimal historicalAverageMm;
+        private Double historicalAverageMm;
         
         /**
          * Projected rainfall deviation (%)
          * Negative = below normal, Positive = above normal
          */
-        private BigDecimal projectedDeviationPercent;
+        private Double projectedDeviationPercent;
         
         /**
          * Scenario type
@@ -125,12 +124,12 @@ public class ClimateRiskDto {
         /**
          * Impact on crop yield (%)
          */
-        private BigDecimal yieldImpactPercent;
+        private Double yieldImpactPercent;
         
         /**
          * Probability of this scenario (%)
          */
-        private BigDecimal probability;
+        private Double probability;
         
         /**
          * Risk level for this scenario
@@ -163,23 +162,23 @@ public class ClimateRiskDto {
         /**
          * Optimal temperature range (°C)
          */
-        private BigDecimal optimalTempMin;
-        private BigDecimal optimalTempMax;
+        private Double optimalTempMin;
+        private Double optimalTempMax;
         
         /**
          * Heat stress threshold (°C)
          */
-        private BigDecimal heatStressThreshold;
+        private Double heatStressThreshold;
         
         /**
          * Cold stress threshold (°C)
          */
-        private BigDecimal coldStressThreshold;
+        private Double coldStressThreshold;
         
         /**
          * Projected temperature deviation (°C)
          */
-        private BigDecimal projectedDeviation;
+        private Double projectedDeviation;
         
         /**
          * Number of extreme heat days projected
