+++ indian-farmer-assistance-app/backend/mandi-service/src/main/java/com/farmer/mandi/dto/MandiLocationDto.java
@@ -1,7 +1,7 @@
 package com.farmer.mandi.dto;
 
 import lombok.*;
-import java.math.BigDecimal;
+
 
 /**
  * DTO for mandi location response.
@@ -21,10 +21,10 @@ public class MandiLocationDto {
     private String state;
     private String district;
     private String address;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private Double latitude;
+    private Double longitude;
     private String contactNumber;
     private String operatingHours;
-    private BigDecimal distanceKm; // Distance from farmer's location
+    private Double distanceKm; // Distance from farmer's location
     private Boolean isActive;
 }
