+++ indian-farmer-assistance-app/backend/mandi-service/src/main/java/com/farmer/mandi/dto/MandiPriceDto.java
@@ -1,7 +1,7 @@
 package com.farmer.mandi.dto;
 
 import lombok.*;
-import java.math.BigDecimal;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -25,13 +25,13 @@ public class MandiPriceDto {
     private String state;
     private String district;
     private LocalDate priceDate;
-    private BigDecimal modalPrice;
-    private BigDecimal minPrice;
-    private BigDecimal maxPrice;
-    private BigDecimal arrivalQuantityQuintals;
+    private Double modalPrice;
+    private Double minPrice;
+    private Double maxPrice;
+    private Double arrivalQuantityQuintals;
     private String unit;
     private String source;
     private LocalDateTime fetchedAt;
-    private BigDecimal distanceKm; // Distance from farmer's location
+    private Double distanceKm; // Distance from farmer's location
     private Boolean isCached; // Indicates if data is from cache
 }
